// src/types/invoices/model/payslip.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zero consult invoicing - OpenAPI 3.2
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Allowance } from './allowance';

export interface Payslip {
    'allowances': Array<Allowance>;
    'employeeId': string;
    'grossSalary': number;
    'id'?: string;
    'month': string;
    'payslipFile'?: string;
    'taxRate': number;
}